change type of size/offset from size_t to uint64_t

+++ services/media_engine/plugins/source/http_source/dash/dash_media_downloader.cpp
@@ -1189,7 +1189,7 @@ void DashMediaDownloader::GetPlaybackInfo(PlaybackInfo& playbackInfo)
     playbackInfo.isDownloading = DownloadFinishStateTmp ? false : true;
 }
 
-size_t DashMediaDownloader::GetBufferSize() const
+uint64_t DashMediaDownloader::GetBufferSize() const
 {
     std::shared_ptr<DashSegmentDownloader> segmentDownloader =
         GetSegmentDownloaderByType(MediaAVCodec::MediaType::MEDIA_TYPE_VID);
